Move view modifier to separate file
